feat: Add TODOs and display methods
Adds some TODO comments based on the README markdown file (instructions
for the project). Also adds the display method for each of the four
objects. All 3 are blank except for the projects object since this was
previously defined. Moved the display method into the object literal
notation earlier in the code.