feat: Change #mapDiv to #map-div in resumeBuilder.js
For some reason, the id for the map was incorrectly called out as
'mapDiv' so this commit updates it to 'map-div' which is what is used
in the index.html and helper.js files. This adds the map section to the
page, though the map itself is not visible yet.