Tweaks & Spinning Wheels
More formatting and content.

Spun wheels on map reload problem after hidden collapsed map and
resize.
Resize with window open works.  It’s complicated, did a bunch of
research, will need to spend more time on jQuery  and events to solve
this one.

Also not going to try an construct the Publications objects at this
time.  Need to turn in the Project in a few hours to meet deadline.
Conceptually it won’t be a problem, just time consuming.