Hoisting Map Problem Solved?
Seems like my locations are getting seen and spit back by google.
Dugway Proving Grounds is the only one without a place marker on the
map right now.  Maybe because it is a Top Secret Military Facility.

Always I did console.log in three places in the helper file, and it
seems that it is working.  All the locations except Dugway now show on
the map after I commented out the bio location line 141 in helper.js…

The following link really digs into what is going on with this
situation, but a bit too much for me to put my whole head around right
now…

https://discussions.udacity.com/t/how-does-camerons-google-map-code-work
/30935