Changed "bioPic" to "biopic"
Required by project evaluator, don’t agree, but made change as required:

“The objects that you create need to follow the names within the schema
exactly. The property name for bio picture should be biopic instead of
bioPic.”

“bioPic” maintains my Javascript Schema in the way the second word
letter is capitalized, and the way that it is already used in
“HTMLbioPic.”  Also “biopic” is a class name in CSS that I had wished
to remain distinguished.  I will contact the grader in email to see if
I am missing something here.